RC 1.0.0

+++ src/ch/sylvainmuller/controller/BookController.java
@@ -4,6 +4,7 @@ import ch.sylvainmuller.models.Book;
 import ch.sylvainmuller.services.BookServiceIt;
 
 import javax.ejb.EJB;
+import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -12,27 +13,24 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(urlPatterns = "/books")
+@WebServlet(urlPatterns = "/books", loadOnStartup = 1)
 public class BookController extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     @EJB
     private BookServiceIt bookService;
 
-    private int cptInit = 0;
+    public void init(ServletConfig config) {
+        bookService.init();
+        System.out.println("Data initialized");
+    }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        /** Initialize data*/
-        if (cptInit == 0 && true == bookService.init()) {
-            cptInit++;
-            System.out.println("Data initialized");
-        }
-
         /** Get id of item to delete */
         String param = request.getParameter("id");
-        if (param != null) {
+        if (param != null && !param.isEmpty()) {
             int id = Integer.parseInt(param);
             bookService.deleteBook(id);
         }
@@ -42,7 +40,7 @@ public class BookController extends HttpServlet {
 
         /** Set attribute and forward */
         request.setAttribute("books", books);
-        request.getRequestDispatcher("book.jsp").forward(request, response);
+        request.getRequestDispatcher("/resources/book.jsp").forward(request, response);
     }
